--- docker-compose.yml
@@ -1,24 +1,0 @@
-version: "3"
-   
-services:
-  xsspot:
-    build: .
-    image: ubt1804:go1.14
-    volumes:
-      - ./xsspot:/root/xsspot
-      - ./xssdata:/root/xssdata
-    working_dir: /root/xsspot
-    command: "/bin/bash -c \"chmod +x ./run.sh && ./run.sh\""
-    ports:
-      - "80:80"
-  xssapi:
-    image: ubt1804:go1.14
-    volumes:
-      - ./xssapi:/root/xssapi
-      - ./xssdata:/root/xssdata
-    working_dir: /root/xssapi
-    command: "/bin/bash -c \"chmod +x ./build_api.sh && ./build_api.sh && npm install && node ./index.js\""
-    ports:
-      - "8099:8099"
-    depends_on:
-      - xsspot
