+++ xss/go.mod
@@ -1,4 +1,4 @@
-module xs
+module xss
 
 go 1.14
