+++ xss/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"xss/server"
+)
+
+func main() {
+	server.Run()
+}
