--- xssapi/build_api.sh
@@ -1,1 +1,0 @@
-npm init -f && npm install express zazler
