--- xsspot/run.sh
@@ -1,1 +1,0 @@
-go build && ./xs
